Make unit tests run in parallel and repeatable.

+++ components/entities/src/main/kotlin/io/collective/entities/OrgIPRecord.kt
@@ -2,4 +2,4 @@ package io.collective.entities
 
 import java.math.BigInteger
 
-data class OrgIPRecord(val id: Long, val startIP: BigInteger, val endIP: BigInteger, val orgId: Int)
+data class OrgIPRecord(val id: Long, val startIP: BigInteger, val endIP: BigInteger, val orgId: Long)
